Add functional tests
