+++ behat.yml.dist
@@ -1,0 +1,24 @@
+default:
+  suites:
+    default:
+      contexts:
+        - 'Behatch\Context\RestContext'
+        - 'Behatch\Context\JsonContext'
+        - 'Behat\MinkExtension\Context\MinkContext'
+  extensions:
+    'Behatch\Extension': ~
+    'Behat\MinkExtension':
+      base_url: 'http://example.com/'
+      files_path: 'features/files'
+      sessions:
+        default:
+          symfony: ~
+    'FriendsOfBehat\SymfonyExtension':
+      bootstrap: features/app/bootstrap.php
+      kernel:
+        environment: test
+        debug: false
+        class: Kernel
+        path: features/app/Kernel.php
+  gherkin:
+    cache: ~
