+++ composer.json
@@ -15,17 +15,12 @@
     "support": {
         "email": "vincentchalamon+support-navision@gmail.com"
     },
-    "repositories": [
-        {
-            "type": "vcs",
-            "url": "https://github.com/vincentchalamon/ntlm-soap-client"
-        }
-    ],
     "require": {
         "php": "^7.2",
         "ext-curl": "*",
         "ext-dom": "*",
         "ext-soap": "*",
+        "doctrine/annotations": "^1.7",
         "doctrine/persistence": "^1.3",
         "doctrine/collections": "^1.2",
         "jamesiarmes/php-ntlm": "^1.0",
@@ -39,8 +34,16 @@
         "symfony/stopwatch": "^4.4 || ^5.0"
     },
     "require-dev": {
+        "ext-simplexml": "*",
         "api-platform/core": "^2.5",
+        "behat/behat": "^3.1",
+        "behat/mink": "^1.7",
+        "behat/mink-browserkit-driver": "^1.3",
+        "behat/mink-extension": "^2.2",
+        "behatch/contexts": "^3.3",
         "easycorp/easyadmin-bundle": "^2.3",
+        "fig/link-util": "^1.1",
+        "friends-of-behat/symfony-extension": "^2.0.10@beta",
         "friendsofphp/php-cs-fixer": "^2.16.1",
         "nyholm/symfony-bundle-test": "^1.6.1",
         "pagerfanta/pagerfanta": "^2.0",
@@ -52,7 +55,8 @@
         "symfony/console": "^4.4 || ^5.0",
         "symfony/dotenv": "^4.4 || ^5.0",
         "symfony/twig-bundle": "^4.4 || ^5.0",
-        "symfony/web-profiler-bundle": "^4.4 || ^5.0"
+        "symfony/web-profiler-bundle": "^4.4 || ^5.0",
+        "zendframework/zend-code": "^3.4"
     },
     "suggest": {
         "api-platform/core": "Add support for API Platform",
@@ -69,11 +73,16 @@
     "autoload-dev": {
         "psr-4": {
             "NavBundle\\Tests\\": "tests/",
-            "NavBundle\\E2e\\": "features/"
+            "NavBundle\\App\\": "features/app/src/"
         }
     },
     "config": {
         "sort-packages": true
     },
-    "minimum-stability": "stable"
+    "minimum-stability": "stable",
+    "extra": {
+        "branch-alias": {
+            "dev-master": "1.0.x-dev"
+        }
+    }
 }
