--- features/TestBundle/Resources/config/services.yaml
@@ -1,13 +1,0 @@
-services:
-  _defaults:
-    autowire: true
-    autoconfigure: true
-    public: false
-
-  NavBundle\E2e\TestBundle\:
-    resource: '../../*'
-    exclude: '../../{DependencyInjection,TestBundle.php}'
-
-  NavBundle\E2e\TestBundle\Controller\:
-    resource: '../../Controller/*'
-    tags: ['controller.service_arguments']
