+++ features/app/mock/responses/INTWS_002_CONT/c22e93460427078c3b11fbb14982543bc39edad9.xml
@@ -1,0 +1,1 @@
+<Soap:Envelope xmlns:Soap="http://schemas.xmlsoap.org/soap/envelope/"><Soap:Body><Read_Result xmlns="urn:microsoft-dynamics-schemas/page/intws_002_cont"><INTWS_002_CONT><Key>32;uhMAAAJ7/0MAQwAwADEAMAAzADEAMw==9;1366849200;</Key><No>CC010313</No><Pr_Mobile_Phone_No>+32</Pr_Mobile_Phone_No><Mobile_Phone_No>14 62 26 52 </Mobile_Phone_No><Name>Gilberte Wynants</Name><Address>Lilsedijk 19</Address><Post_Code>B-2340</Post_Code><City>BEERSE</City><Phone_No>A renseigner</Phone_No><Salesperson_Code>BF</Salesperson_Code><Country_Region_Code>BE</Country_Region_Code><E_Mail>n.grigorova@groupe-hli.com</E_Mail><Home_Page>FOURNISSEUR</Home_Page><Type>Person</Type><Company_No>CO001561</Company_No><Company_Name>SUEZ</Company_Name><Type_Tiers>Vendor</Type_Tiers><No_Tiers>00000437</No_Tiers><Portail_Y_N>true</Portail_Y_N></INTWS_002_CONT></Read_Result></Soap:Body></Soap:Envelope>
