+++ features/app/mock/responses/INTWS_002_CONT/cff06140266f9a21e606832c03c9128bc862ca50.xml
@@ -1,0 +1,1 @@
+<Soap:Envelope xmlns:Soap="http://schemas.xmlsoap.org/soap/envelope/"><Soap:Body><Update_Result xmlns="urn:microsoft-dynamics-schemas/page/intws_002_cont"><INTWS_002_CONT><Key>32;uhMAAAJ7/0MAQwAwADAAOQA4ADkANA==9;1368091290;</Key><No>CC009894</No><Name>Jane DOE</Name><Address>34-36 rue Alphonse Pluchet</Address><Address_2>BP 200</Address_2><Post_Code>92225</Post_Code><City>BAGNEUX CEDEX</City><Phone_No>A renseigner</Phone_No><Country_Region_Code>FR</Country_Region_Code><E_Mail>jane.doe@example.com</E_Mail><Home_Page>FOURNISSEUR</Home_Page><Type>Person</Type><Company_No>CO001142</Company_No><Company_Name>DEKRA INSPECTION</Company_Name><Type_Tiers>Vendor</Type_Tiers><No_Tiers>00000002</No_Tiers><Portail_Y_N>true</Portail_Y_N></INTWS_002_CONT></Update_Result></Soap:Body></Soap:Envelope>
