+++ features/app/mock/responses/INTWS_020_DI_HEADER/1e1e4c3bddcf01d51e747b56bc337f99c2093b2d.xml
@@ -1,0 +1,1 @@
+<Soap:Envelope xmlns:Soap="http://schemas.xmlsoap.org/soap/envelope/"><Soap:Body><Update_Result xmlns="urn:microsoft-dynamics-schemas/page/intws_020_di_header"><INTWS_020_DI_HEADER><Key>36;c8MAAAJ7/0kAUgBDAE8AMAAwADAAMAAyADk=9;1368090930;</Key><No>IRCO000029</No><Order_Date>2020-03-11</Order_Date><DI_contact>CO010358</DI_contact><Status_IR>Email</Status_IR><Asset_Manager>BF</Asset_Manager><Building_Code>BRUXELLE</Building_Code><Address>Kleine Kloosterstart 10</Address><Post_Code>1932</Post_Code><City>ZAVENTERN</City><comment>Test NavBundle</comment><Appointment_Date>0001-01-01</Appointment_Date><Appointment_Time>00:00:00.0000000</Appointment_Time><Create_SO>0001-01-01</Create_SO><URL_DI_du_portail>https://pp.corum-tenant-support.com/intervention/IRCO000029</URL_DI_du_portail><Customer_No>BRIDGEST</Customer_No><Customer_Name>Bridgestone Europe SA</Customer_Name><Vendor_No>00000437</Vendor_No><Included_in_contract>Included</Included_in_contract><Service_No>078</Service_No><Service_Description>Recycling</Service_Description><Vendor_Contract_No>VC000006</Vendor_Contract_No><Vendor_Name>SUEZ</Vendor_Name><Vendor_Address>Lilsedijk 19</Vendor_Address><Vendor_Post_Code>B-2340</Vendor_Post_Code><Vendor_City>BEERSE</Vendor_City><Vendor_Country_Code>BE</Vendor_Country_Code><Vendor_Contact>CC010313</Vendor_Contact><Vendor_Contact_Name>Gilberte Wynants</Vendor_Contact_Name><Vendor_Contact_E_mail>n.grigorova@groupe-hli.com</Vendor_Contact_E_mail><Send_Mail>2020-03-11T10:20:57.237Z</Send_Mail><CGV_Accepted>false</CGV_Accepted></INTWS_020_DI_HEADER></Update_Result></Soap:Body></Soap:Envelope>
