+++ features/app/mock/responses/INTWS_020_DI_HEADER/a8b0cb96137d7d067f1af07253c69a30135d3d29.xml
@@ -1,0 +1,1 @@
+<Soap:Envelope xmlns:Soap="http://schemas.xmlsoap.org/soap/envelope/"><Soap:Body><Create_Result xmlns="urn:microsoft-dynamics-schemas/page/intws_020_di_header"><INTWS_020_DI_HEADER><Key>36;c8MAAAJ7/0kAUgBDAE8AMAAwADAAMAA1ADY=9;1368090020;</Key><No>IRCO000056</No><Order_Date>2020-03-27</Order_Date><Status_IR>Stand_By_Validation</Status_IR><Asset_Manager>BF</Asset_Manager><Building_Code>BRUXELLE</Building_Code><Address>Kleine Kloosterstart 10</Address><Post_Code>1932</Post_Code><City>ZAVENTERN</City><Country_Code>BE</Country_Code><Appointment_Date>0001-01-01</Appointment_Date><Appointment_Time>00:00:00.0000000</Appointment_Time><Create_SO>0001-01-01</Create_SO><URL_DI_du_portail>https://pp.corum-tenant-support.com/intervention/IRCO000056</URL_DI_du_portail><Customer_No>BRIDGEST</Customer_No><Customer_Name>Bridgestone Europe SA</Customer_Name><Vendor_No>00000437</Vendor_No><Included_in_contract>_blank__blank_</Included_in_contract><Service_No>078</Service_No><Service_Description>Recycling</Service_Description><Vendor_Contract_No>VC000006</Vendor_Contract_No><Vendor_Name>SUEZ</Vendor_Name><Vendor_Address>Lilsedijk 19</Vendor_Address><Vendor_Post_Code>B-2340</Vendor_Post_Code><Vendor_City>BEERSE</Vendor_City><Vendor_Country_Code>BE</Vendor_Country_Code><Vendor_Contact>CC010313</Vendor_Contact><Vendor_Contact_Name>Gilberte Wynants</Vendor_Contact_Name><Vendor_Contact_E_mail>n.grigorova@groupe-hli.com</Vendor_Contact_E_mail><Send_Mail>0001-01-01T00:00:00</Send_Mail><CGV_Accepted>false</CGV_Accepted></INTWS_020_DI_HEADER></Create_Result></Soap:Body></Soap:Envelope>
