--- features/bootstrap.php
@@ -1,39 +1,0 @@
-<?php
-
-/*
- * This file is part of the NavBundle.
- *
- * (c) Vincent Chalamon <vincentchalamon@gmail.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-use Symfony\Component\Dotenv\Dotenv;
-
-date_default_timezone_set('UTC');
-
-require __DIR__.'/../vendor/autoload.php';
-require __DIR__.'/Kernel.php';
-
-// Load cached env vars if the .env.local.php file exists
-// Run "composer dump-env prod" to create it (requires symfony/flex >=1.2)
-if (is_array($env = @include __DIR__.'/.env.local.php')) {
-    foreach ($env as $k => $v) {
-        $_ENV[$k] = $_ENV[$k] ?? (isset($_SERVER[$k]) && 0 !== strpos($k, 'HTTP_') ? $_SERVER[$k] : $v);
-    }
-} else {
-    // load all the .env files
-    if (method_exists(Dotenv::class, 'loadEnv')) {
-        (new Dotenv(false))->loadEnv(__DIR__.'/.env');
-    } else {
-        (new Dotenv())->load(__DIR__.'/.env');
-    }
-}
-
-$_SERVER += $_ENV;
-$_SERVER['APP_ENV'] = $_ENV['APP_ENV'] = ($_SERVER['APP_ENV'] ?? $_ENV['APP_ENV'] ?? null) ?: 'dev';
-$_SERVER['APP_DEBUG'] = $_SERVER['APP_DEBUG'] ?? $_ENV['APP_DEBUG'] ?? 'prod' !== $_SERVER['APP_ENV'];
-$_SERVER['APP_DEBUG'] = $_ENV['APP_DEBUG'] = (int) $_SERVER['APP_DEBUG'] || filter_var($_SERVER['APP_DEBUG'], FILTER_VALIDATE_BOOLEAN) ? '1' : '0';
